requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
idna==3.11
praw==7.8.1
prawcore==2.4.0
pymongo==4.15.4
PyYAML==6.0.3
requests==2.32.5
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.9.0
